Split state into feature-specific files
