+++ lib/screens/home/screen.dart
@@ -1,5 +1,7 @@
 import 'package:concordium_wallet/screens/terms_and_conditions/screen.dart';
-import 'package:concordium_wallet/state.dart';
+import 'package:concordium_wallet/state/network.dart';
+import 'package:concordium_wallet/state/services.dart';
+import 'package:concordium_wallet/state/terms_and_conditions.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -16,7 +18,7 @@ class _HomeScreenState extends State<HomeScreen> {
   void initState() {
     super.initState();
 
-    // Fetch currently valid T&C version.
+    // Fetch currently valid T&C version when initializing the widget.
     // TODO: Use the "last verified" time to determine whether to perform the refresh now and whenever the home screen is loaded.
     final network = context.read<SelectedNetwork>().state;
     final services = context.read<ServiceRepository>().networkServices[network]!;
@@ -36,7 +38,14 @@ class _HomeScreenState extends State<HomeScreen> {
             if (validTac == null) {
               // Show spinner if no valid T&C have been resolved yet (not as a result of actually ongoing fetch).
               return const Center(
-                child: CircularProgressIndicator(),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    CircularProgressIndicator(),
+                    Text('Waiting for Terms & Conditions...'),
+                  ],
+                ),
               );
             }
             final acceptedTac = tacState.accepted;
@@ -57,13 +66,24 @@ class _HomeScreenState extends State<HomeScreen> {
                   ),
                 ),
                 ElevatedButton(
-                  onPressed: () => context.read<TermsAndConditionAcceptance>().resetAcceptedTime(),
-                  child: const Text('Reset check time'),
+                  onPressed: () => context.read<TermsAndConditionAcceptance>().resetAccepted(),
+                  child: const Text('Reset accepted T&C'),
                 ),
                 const SizedBox(height: 8),
                 ElevatedButton(
-                  onPressed: () => context.read<TermsAndConditionAcceptance>().resetAcceptedVersion(),
-                  child: const Text('Reset accepted version'),
+                  onPressed: () => context.read<TermsAndConditionAcceptance>().resetValid(),
+                  // NOTE: This breaks the app because no re-fetch is triggered - hot restart is necessary to recover.
+                  child: const Text('Reset valid T&C (breaks the app!)'),
+                ),
+                const SizedBox(height: 8),
+                ElevatedButton(
+                  onPressed: () => context.read<TermsAndConditionAcceptance>().testSetAcceptedVersion('1.2.3'),
+                  child: const Text('Set accepted version to "1.2.3"'),
+                ),
+                const SizedBox(height: 8),
+                ElevatedButton(
+                  onPressed: () => context.read<TermsAndConditionAcceptance>().testResetAcceptedTime(),
+                  child: const Text('Reset accepted time'),
                 ),
               ],
             );
